==> src/02_data_cleaning.ipynb <==
# 02. Data Cleaning

import pandas as pd
import glob
import re
import os

## Read the 70 csv files together

news_page_all = glob.glob("../data/raw/*.csv")

# news_page_all

df_list = []

for page in news_page_all:
    # print(page)
    df = pd.read_csv(page)
    df_list.append(df)

news_dataset = pd.concat(df_list, ignore_index=False)

len(news_dataset)

news_dataset.describe()

news_dataset.head()

news_dataset.columns

news_dataset.shape

news_dataset.size

## Data Cleaning

### Removing duplicate values

news_dataset = news_dataset.drop_duplicates()
news_dataset

### Drop missing values

news_dataset = news_dataset.dropna()
news_dataset

news_dataset["Article"]

### Reduces unnessary information `news_dataset['Article']`

news_dataset["Article"].iloc[0]

pattern = r"Home>>\n\t\t\n.*?\d{1,2}:\d{2}, [A-Za-z]+ \d{1,2}, \d{4}"
news_dataset['Article'].str.contains(pattern, regex=True).sum()
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article']

pattern = r"\(Photo: .*?\)"
news_dataset['Article'].str.contains(pattern, regex=True).sum()
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"\(Photo: .*?\)"
news_dataset['Article'].str.contains(pattern, regex=True).sum()
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"\(Web editor: .*?\)"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"Co."
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"Ltd."
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"'s"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"\n"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"\(Photo courtesy .*?\)"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "__replace__", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"__replace__"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(r'[^\w\s_\-\--]', ' ', x))
news_dataset['Article']

news_dataset['Article'].iloc[36]

pattern = r"--"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Article'].iloc[0]

pattern = r"U.S"
print(news_dataset['Article'].str.contains(pattern, regex=True).sum())
news_dataset['Article'] = news_dataset['Article'].apply(lambda x: re.sub(pattern, "USA", x, flags=re.DOTALL))
news_dataset['Article'].iloc[198]

print(news_dataset['Article'].iloc[0])
print(news_dataset['Article'].iloc[1])
print(news_dataset['Article'].iloc[2])
print(news_dataset['Article'].iloc[3])
print(news_dataset['Article'].iloc[4])

### Reduces unnessary information `news_dataset['Title']`

news_dataset['Title']

print(news_dataset['Title'].iloc[0])
print(news_dataset['Title'].iloc[1])
print(news_dataset['Title'].iloc[2])
print(news_dataset['Title'].iloc[3])
print(news_dataset['Title'].iloc[4])

for i in range(0,10):
    print(i, news_dataset['Title'].iloc[i])

pattern = r"'s"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title'].iloc[0]

pattern = r"'e"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title'].iloc[198]

pattern = r"-"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title'].iloc[0]

news_dataset['Title'].iloc[198]

pattern = r"U.S"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, "USA", x, flags=re.DOTALL))
news_dataset['Title'].iloc[198]

pattern = r"Chang’e"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title'].iloc[198]

pattern = r":"
print(news_dataset['Title'].str.contains(pattern, regex=True).sum())
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title'].iloc[198]

pattern = r"\""
print(news_dataset['Title'][news_dataset['Title'].str.contains(pattern, regex=True)])
news_dataset['Title'] = news_dataset['Title'].apply(lambda x: re.sub(pattern, " ", x, flags=re.DOTALL))
news_dataset['Title']

### Remove Null rows

news_dataset = news_dataset[news_dataset['Article'].str.len() >= 10]

news_dataset

## Save as `clean/news_dataset.csv`

filename = '../data/clean/news_dataset.csv'
if os.path.exists(filename):
    response = input(f"The file {filename} already exists. Do you want to overwrite it? (y/n): ")
    if response.lower() == 'y':
        news_dataset.to_csv(filename, index=False)
    else:
        print("File not overwritten.")
else:
    news_dataset.to_csv(filename, index=False)
