==> src/03_data_preprocessing.ipynb <==
# 03.Data Preprocessing

import numpy as np
import pandas as pd
import os

import seaborn as sns
import matplotlib.pyplot as plt

import re
import nltk

news = pd.read_csv('../data/clean/news_dataset.csv')
news.head()

## 1. Lowercasing text

news.columns

news.columns = news.columns.str.lower()

news.head()

news['title'] = news['title'].str.lower()
news['title']

news['article'] = news['article'].str.lower()
news['article']

news['article'].iloc[0]

## 2. Remove stop word

from nltk.corpus import stopwords
nltk.download('stopwords')

stop_words = set(stopwords.words('english'))

def remove_stop_words(text):
    return ' '.join([word for word in text.split() if word.lower() not in stop_words])

news['title'] = news['title'].apply(remove_stop_words)

news['article'] = news['article'].apply(remove_stop_words)

news['title'].iloc[0]

news['article'].iloc[0]

## 3. Remove `numbers`

news['title'] = news['title'].apply(lambda x: re.sub(r'\d+', '', x))

news['article'] = news['article'].apply(lambda x: re.sub(r'\d+', '', x))

news['title'].iloc[0]

news['article'].iloc[0]

## 4. Lemmatization or stemming

from nltk.stem import WordNetLemmatizer
nltk.download('wordnet')

lemmatizer = WordNetLemmatizer()

def lemmatize_text(text):
    words = nltk.word_tokenize(text)
    lemmatized_words = [lemmatizer.lemmatize(word) for word in words]
    return ' '.join(lemmatized_words)

news['title'] = news['title'].apply(lemmatize_text)

news['article'] = news['article'].apply(lematize_text)

news['title'].iloc[0]

news['article'].iloc[0]

## 5. Handling contractions

import contractions

def expand_contractions(text):
    return contractions.fix(text)

news['title'] = news['title'].apply(expand_contractions)

news['article'] = news['article'].apply(expand_contractions)

news['title'].iloc[0]

news['article'].iloc[0]

## 6. Removing extra whitespaces

news['title'] = news['title'].apply(lambda x: x.strip())

news['article'] = news['article'].apply(lambda x: x.strip())

news['title'].iloc[0]

news['article'].iloc[0]

## Save as `preprocessed/news_preprocessed.csv`

news.to_csv('../data/preprocessed/news_preprocessed.csv', index=False)
