==> src/04_news_columns_generator.ipynb <==
# 04. News Columns Generator

在新闻网站 (http://en.people.cn/) 上，我很遗憾没有找到预期的`栏目`，只留下了8个栏目（滚动、独家、科学家、产业动态、发明·创新、探索·发现、医学·健康、航)。因此，我使用预先训练的零样本分类模型对文章进行分类。我已将结果储存为 `news['news_column']`。

# !pip install transformers

import pandas as pd

news = pd.read_csv('../data/preprocessed/news_preprocessed.csv')
news.head()

## Define `classifier`

from transformers import pipeline

classifier = pipeline(
    "zero-shot-classification",
    model="facebook/bart-large-mnli",
    device=0  
)

candidate_labels = ['rolling', 'exclusive', 'scientist', 'industry dynamics', ' innovation', 'exploration discovery', 'medicine & health', 'aerospace']

## Labeling 

news['news_column'] = news['article'].apply(lambda x: classifier(x, candidate_labels)['labels'][0])

news.head()

news.columns

## Save as `news_labeled.csv`

news.to_csv('../data/labeled/news_labeled.csv', index=False)
